requirements.txt
python-telegram-bot==21.10
anthropic==0.52.0
aiohttp==3.11.14
python-dotenv==1.1.0
